+++ src/modmux/__init__.py
@@ -14,10 +14,17 @@ from .models import (
     Provider,
     ProviderCreds,
 )
+from .providers.curseforge import CurseforgeCreds
+from .providers.modio import ModioCreds
+from .providers.modrinth import ModrinthCreds
+from .providers.nexusmods import NexusCreds
+from .providers.steam import SteamCreds
+from .providers.wube import WubeCreds
 from .utils.urls import parse_url
 
 __all__ = [
     "Author",
+    "CurseforgeCreds",
     "Dependency",
     "FileAsset",
     "LocaleTag",
@@ -26,9 +33,14 @@ __all__ = [
     "ModID",
     "ModSummary",
     "ModVersion",
+    "ModioCreds",
+    "ModrinthCreds",
     "Muxer",
+    "NexusCreds",
     "Provider",
     "ProviderCreds",
+    "SteamCreds",
     "modmux_client",
     "parse_url",
+    "WubeCreds",
 ]